requirements.txt
streamlit
streamlit-webrtc
opencv-python
mediapipe
torch
torchvision
numpy<2
Pillow
